Fix for #3587919, plus general test cleanup

+++ sveditor/plugins/net.sf.sveditor.core.tests/src/net/sf/sveditor/core/tests/content_assist/TestContentAssistBuiltins.java
@@ -72,7 +72,10 @@ public class TestContentAssistBuiltins extends TestCase {
 		super.tearDown();
 
 		fIndexRgy.save_state();
-		TestUtils.delete(fTmpDir);
+		
+		if (fTmpDir != null && fTmpDir.exists()) {
+			TestUtils.delete(fTmpDir);
+		}
 	}
 
 	public void testCovergroupOption() {
